timezonechange/worker-timezonechange.html

<!DOCTYPE html>
<title>Test worker handle ontimezonechange event.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
promise_test(async t => {
  const oldTimeZone = "Pacific/Fakaofo";
  const newTimeZone = "Asia/Taipei";
  // First we set the time zone to a fixed time zone.
  await window.test_driver.set_time_zone(oldTimeZone);
  const worker = new Worker('resources/worker-timezonechange.js');
  return new Promise(r => { worker.onmessage = r; })
    .then(e => {
      // Once we know the worker is ready, we change the timezone.
      assert_equals(e.data, "READY:" + oldTimeZone);
      return new Promise(async r => {
        worker.onmessage = r;
        await window.test_driver.set_time_zone(newTimeZone);
      }).then(e => {
        assert_equals(e.data, "SUCCESS:" + newTimeZone);
      })
    });
}, "Test a dedicated worker handles ontimezonechange event.");
</script>
